moved model config to superclass
