+++ fhirzeug/generators/python_pydantic/templates/resource_header.py
@@ -1,7 +1,6 @@
 import typing
 import enum
 import pydantic
-from collections.abc import Mapping
 
 
 def choice_of_validator(choices, optional):
@@ -17,20 +16,3 @@ def choice_of_validator(choices, optional):
         return values
 
     return check_at_least_one
-
-
-def camelcase_alias_generator(name: str) -> str:
-    """Maps snakecase to camelcase.
-    
-    This enables members to be created from camelCase. It takes the existing camelcase membername
-    like foo_bar and converts it to its camelcase pedant fooBar.
-    
-    Additionally it removes trailing _, since this is used to make membernames of reserved keywords
-    usable, like `class`.
-    """
-
-    if name.endswith("_"):
-        return name[:-1]
-
-    components = name.split("_")
-    return components[0] + "".join(word.capitalize() for word in components[1:])
